backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
python-multipart>=0.0.6
alembic>=1.12.0
